requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
Django==3.1.2
django-rename-app==0.1.2
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.5.0
djoser==2.0.5
filelock==3.0.12
idna==2.10
importlib-metadata==2.0.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.34
virtualenvwrapper-win==1.2.6
zipp==3.3.0
